% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listview.R
\name{listview}
\alias{listview}
\title{View object as collapsible nested list in HTML}
\usage{
listview(x)
}
\arguments{
\item{x}{An atomic vector or a list}
}
\value{
A \code{list()} with a \code{shiny.tag} class that can be converted into an
HTML string via \code{as.character()} and saved to a file with \code{save_html()}.
}
\description{
View object as collapsible nested list in HTML
}
\examples{
x <- list(
  list(id = "a", val = 2),
  list(
    id = "b",
    val = 1,
    children = list(
      list(id = "b1", val = 2.5),
      list(
        id = "b2",
        val = 8,
        children = list(
          list(id = "b21", val = 4)
        )
      )
    )
  ),
  list(
    id = "c",
    val = 8,
    children = list(
      list(id = "c1"),
      list(id = "c2", val = 1)
    )
  )
)

listview(x)
}
